additional port
A while ago, I ported this QuadTree implementation to Kit. It's a very different language (and in alpha development), but my port works just about the same with the same methods (and a few extra). Feel free to not accept the PR if you don't think mine should be listed. [More info about Kit here.](https://www.kitlang.org)